Move dependencies to to common.h
